Add files via upload
"Excited to share my recent coding project: a Java solution for a problem-solving challenge! 🚀✨ I have implemented an algorithm that calculates the next number in a series based on the given initial values. The code utilizes a Scanner to receive user inputs and follows a modular approach to ensure efficient and reusable code.

I have completed this project as part of my ongoing coding journey and participation in a 100 Days Coding Challenge. If you're interested in exploring the code or want to see my progress throughout the challenge, you can find the complete Java implementation in my GitHub repository below.

Feel free to check it out, provide feedback, or even suggest improvements! I believe in continuous learning and growth, and I'm excited to share my coding experiences with the community. Stay tuned for more updates on my coding journey! 🌟💻"